version check does not open a new window but shows in the top message box area. it also does not any more offer to stop the program, you are already in the right place to do so. closes #98
